Depletion Rate Implementation & Bug Fixes
- Depletion Rate Implemented 
- Fixed bug with showing the energy level instead of power level.
- Added as triggered events and not just adding and subtracting power.

+++ Assets/Project Alpha/Common/Scripts/Resource Systems/Power/PowerEvent.cs
@@ -12,9 +12,9 @@ namespace MoreMountains.CorgiEngine
     public struct PowerEvent 
     {
         public PowerEventType powerEventType;
-        public int amount;
+        public float amount;
 
-        public PowerEvent (PowerEventType powerEventType, int amount) {
+        public PowerEvent (PowerEventType powerEventType, float amount) {
             this.powerEventType = powerEventType;
             this.amount = amount;
         }
